commands: move move code to move.rs
